Updated: light folder structure refactoring
