+++ src/parametric/core.py
@@ -1,6 +1,6 @@
 import networkx as nx
 import matplotlib.pyplot as plt
-from discrete_morse_theory import V_paths, critical_cells, gradient
+from discrete.core import V_paths, critical_cells, gradient
 from helpers import get_skel, magic_function
 
 plt.rcParams["figure.figsize"] = (10, 10)
